pyproject.toml
[project]
name = "my-custom-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "strands-agents>=1.5.0",
    "uvicorn[standard]>=0.35.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
]
